+++ common/src/main/resources/flywheel.backend.mixins.json
@@ -8,6 +8,7 @@
     "AbstractClientPlayerAccessor",
     "GameRendererAccessor",
     "GlStateManagerMixin",
+    "LevelRendererAccessor",
     "LightTextureAccessor",
     "OptionsMixin",
     "OverlayTextureAccessor",
