+++ forge/src/main/java/com/jozufozu/flywheel/platform/ClientPlatformImpl.java
@@ -7,10 +7,14 @@ import com.jozufozu.flywheel.api.event.ReloadLevelRendererEvent;
 import com.jozufozu.flywheel.api.event.RenderContext;
 import com.jozufozu.flywheel.api.event.RenderStage;
 import com.jozufozu.flywheel.api.event.RenderStageEvent;
+import com.jozufozu.flywheel.config.FlwConfig;
+import com.jozufozu.flywheel.config.FlwForgeConfig;
 import com.jozufozu.flywheel.lib.util.ShadersModHandler;
 
 import net.irisshaders.iris.api.v0.IrisApi;
 import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.ModList;
 
@@ -56,5 +60,15 @@ public class ClientPlatformImpl extends ClientPlatform {
 		} else {
 			return null;
 		}
+	}
+
+	@Override
+	public int getLightEmission(BlockState state, ClientLevel level, BlockPos pos) {
+		return state.getLightEmission(level, pos);
+	}
+
+	@Override
+	public FlwConfig getConfigInstance() {
+		return FlwForgeConfig.INSTANCE;
 	}
 }
